Rename functions, adjust resulting signed payload
