+++ packages/nouns-webapp/functions/v0-noun-owners/index.ts
@@ -1,12 +1,12 @@
 import { Handler } from '@netlify/functions';
-import { isNounDelegate, normalizeNouns, nounsQuery } from '../../theGraph';
+import { normalizeNouns, nounsQuery } from '../theGraph';
 
 const handler: Handler = async (event, context) => {
   const graphResponse = await nounsQuery();
   const nouns = normalizeNouns(graphResponse.data.data.nouns);
   return {
     statusCode: 200,
-    body: JSON.stringify(isNounDelegate(event.body, nouns)),
+    body: JSON.stringify(nouns),
   };
 };
